requirements.txt
langchain
langchain_community
langchain-tavily
langchain[google-genai]
faiss-cpu
sentence-transformers
python-dotenv
pypdf
tf-keras